+++ src/rabbitizer/src/instr/instr_suffix.rs
@@ -1,6 +1,0 @@
-/* SPDX-FileCopyrightText: © 2024-2025 Decompollaborate */
-/* SPDX-License-Identifier: MIT */
-
-use crate::instr::InstrSuffix;
-
-impl InstrSuffix {}