+++ src/rabbitizer/src/instr/mod.rs
@@ -2,13 +2,11 @@
 /* SPDX-License-Identifier: MIT */
 
 mod instr_field;
-mod instr_suffix;
 mod instr_type;
 mod instruction;
 mod instruction_display;
 mod instruction_flags;
-
-pub use crate::generated::InstrSuffix;
+mod mnemonic_display;
 
 #[allow(deprecated)]
 pub use instr_type::InstrType;
@@ -17,3 +15,4 @@ pub use instr_field::InstrField;
 pub use instruction::Instruction;
 pub use instruction_display::InstructionDisplay;
 pub use instruction_flags::InstructionFlags;
+pub use mnemonic_display::MnemonicDisplay;